% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cognitive-scores.R
\name{score_cdr}
\alias{score_cdr}
\title{Score the Clinical Dementia Rating (CDR)}
\usage{
score_cdr(
  memory,
  orientation,
  judgment,
  community,
  home_hobbies,
  personal_care
)
}
\arguments{
\item{memory}{Memory domain rating (0--3).}

\item{orientation}{Orientation domain rating (0--3).}

\item{judgment}{Judgment and problem solving rating (0--3).}

\item{community}{Community affairs rating (0--3).}

\item{home_hobbies}{Home and hobbies rating (0--3).}

\item{personal_care}{Personal care rating (0--3).}
}
\value{
A list with elements \code{global} (the global CDR score) and
\code{sb} (the sum of boxes).
}
\description{
Computes the global CDR score using the Washington University algorithm
(Morris, 1993) and the sum of boxes (CDR-SB) from domain ratings.
}
